+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Added
 - Bars can now be built by `beginTime` instead of `endTime`
+- Added `DurationBarBuilder` to aggregate Bars after a fixed duration.
 
 
 ## 0.18 (released May 15, 2025)